install a separate syntax table for syntax parsing
keep the sql-mode syntax table unchanged, but use a special one during
syntactic parsing.  This is done so that navigation over syntactic elements is
simpler for parsing, but stays reasonable while the user edits the file.

For example, the string "table.column" will be considered one symbol for
indentation purposes, but two symbols for editing (so an M-b will move to the
dot)