Establish a limit for syntactic information search
See also issue #26

`sqlind-syntax-of-line' will use `sqlind-begining-of-directive' to establish a
search limit for the current syntax.  This avoids calling
`sqlind-begining-of-directive' repeatedly.

This changes the indentation anchor for toplevel statements when directives
are present in the file.  Unit test data updated accordingly.

* sql-indent.el (sqlind-search-limit): new variable
(sqlind-beginning-of-statement-1): avoid searching before
`sqlind-search-limit'
(sqlind-beginning-of-statement): use `sqlind-search-limit'
(sqlind-maybe-loop-statement): fix mispelling of throw symbol
(sqlind-beginning-of-block): limit search to `sqlind-search-limit'
(sqlind-syntax-of-line): bind `sqlind-search-limit' to
`sqlind-begining-of-directive'