Introduce a nested-statement-close syntax
This is recognized inside a nested statement when the line to be indented
starts with a closing bracket.  This is intended to replace
`sqlind-lineup-close-paren-to-open` and
`sqlind-lineup-close-paren-to-open-indentation` indentation helpers which had
to be added to too many syntactic symbols.

With the new syntax symbol, using `sqlind-use-anchor-indentation` will lineup
hanging close brackets with the start of the statement and
`sqlind-use-anchor-indentation` will line them up with the actual opening
bracket itself.