Fix detection of DECLARE blocks in PostgresSQL (#92)
In PostgresSQL, the DECLARE keyword might begin a block of declarations, or
declare a single cursor.  Updated the syntax scanning code to only consider
DECLARE a block start when it is part of a program block.

This change will alter how PostgresSQL blocks are indented, hopefully for the
better.

Added tests for the relevant cases, and updated existing test cases which also
incorrectly detected DECLARE statements.