fixups for indenting where clauses (#93)
A line under the "where" clause should be lined up with the first element
after the where clause, unless it is a logical operator, in which case the
operator should line up under the where clause. Fixed how the base column is
calculated in sqlind-left-justify-logical-operator.

Some tests (pr17 and p19) had to be fixed as a result. Also introduced pr93 to
verify this test case.